DRY context menu wrapper for segment and skeleton tab
